refactor(jina): 리렌더링 최적화를 위해 zustand 적용 #61
